refactor: 重构useZen，用useAllProgress解耦
