+++ hooks/useZen.ts
@@ -1,8 +1,6 @@
 "use client";
 
-import { Progress } from "@hooks/useProgress";
 import { useSuspenseQuery } from "@tanstack/react-query";
-import { useEffect } from "react";
 
 // Question Data Type
 interface ConstQuestion {
@@ -16,14 +14,7 @@ interface ConstQuestion {
   _hash: number;
 }
 
-const LC_RATING_PROGRESS_KEY = (questionID: string) =>
-  `lc-rating-zen-progress-${questionID}`;
-const LC_RATING_ZEN_LAST_USED_FILTER_KEY = `lc-rating-zen-last-used-filter`;
-
-// Progress Related
-type ProgressData = Record<string, string>;
-
-export function useZen(setLocalStorageProgressData: any) {
+export function useZen() {
   const { data, isFetching } = useSuspenseQuery({
     queryKey: [],
     queryFn: () =>
@@ -34,16 +25,6 @@ export function useZen(setLocalStorageProgressData: any) {
         }),
     refetchOnWindowFocus: false,
   });
-  useEffect(() => {
-    const loadedLocalStorageData: ProgressData = {};
-    data.forEach((item) => {
-      loadedLocalStorageData[item.question_id] =
-        localStorage.getItem(LC_RATING_PROGRESS_KEY(item.question_id)) ||
-        Progress.TODO;
-    });
-    if (setLocalStorageProgressData) {
-      setLocalStorageProgressData(loadedLocalStorageData);
-    }
-  }, [data]);
+
   return { zen: data, isPending: isFetching };
 }
